# apps/contracts/openapi/cambot-api.yaml
openapi: 3.1.0
info:
  title: CamBot REST API
  version: 0.1.0
  description: Product API used by the CamBot client.
servers:
  - url: /api
    description: Reverse-proxied CamBot API root
security:
  - basicAuth: []
...
paths:
  /camera-groups:
    get:
      tags: [Camera Groups]
      summary: List CamBot camera groups
      operationId: listCameraGroups
      responses:
        "200":
          description: Camera groups returned
          content:
            application/json:
              schema:
                $ref: "./schemas/cambot/camera-group.yaml#/components/schemas/CameraGroupListResponse"

    post:
      tags: [Camera Groups]
      summary: Create a CamBot camera group
      operationId: createCameraGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/cambot/camera-group.yaml#/components/schemas/CreateCameraGroupRequest"
      responses:
        "201":
          description: Camera group created
          content:
            application/json:
              schema:
                $ref: "./schemas/cambot/camera-group.yaml#/components/schemas/CameraGroup"
  /camera-groups/{groupId}:
    get:
      tags: [Camera Groups]
      summary: Get a CamBot camera group
      operationId: getCameraGroup
      parameters:
        - $ref: "#/components/parameters/GroupId"
      responses:
        "200":
          description: Camera group returned
          content:
            application/json:
              schema:
                $ref: "./schemas/cambot/camera-group.yaml#/components/schemas/CameraGroup"

    put:
      tags: [Camera Groups]
      summary: Update a CamBot camera group
      operationId: updateCameraGroup
      parameters:
        - $ref: "#/components/parameters/GroupId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/cambot/camera-group.yaml#/components/schemas/UpdateCameraGroupRequest"
      responses:
        "200":
          description: Camera group updated
          content:
            application/json:
              schema:
                $ref: "./schemas/cambot/camera-group.yaml#/components/schemas/CameraGroup"

    delete:
      tags: [Camera Groups]
      summary: Delete a CamBot camera group
      operationId: deleteCameraGroup
      parameters:
        - $ref: "#/components/parameters/GroupId"
      responses:
        "204":
          description: Camera group deleted
  /camera-groups/{groupId}/cameras:
    put:
      tags: [Camera Groups]
      summary: Replace cameras assigned to a CamBot camera group
      operationId: replaceCameraGroupCameras
      parameters:
        - $ref: "#/components/parameters/GroupId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/cambot/camera-group.yaml#/components/schemas/ReplaceCameraGroupCamerasRequest"
      responses:
        "200":
          description: Camera group cameras replaced
          content:
            application/json:
              schema:
                $ref: "./schemas/cambot/camera-group.yaml#/components/schemas/CameraGroup"
  /camera-groups/{groupId}/stats:
    get:
      tags: [Camera Groups]
      summary: Get camera group statistics
      operationId: getCameraGroupStats
      parameters:
        - $ref: "#/components/parameters/GroupId"
      responses:
        "200":
          description: Camera group statistics returned
          content:
            application/json:
              schema:
                $ref: "./schemas/cambot/camera-group.yaml#/components/schemas/CameraGroupStats"
  /saved-prompts:
    get:
      tags: [Saved Prompts]
      summary: List saved prompts
      operationId: listSavedPrompts
      responses:
        "200":
          description: Saved prompts returned
          content:
            application/json:
              schema:
                $ref: "./schemas/cambot/saved-prompt.yaml#/components/schemas/SavedPromptListResponse"

    post:
      tags: [Saved Prompts]
      summary: Create a saved prompt
      operationId: createSavedPrompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/cambot/saved-prompt.yaml#/components/schemas/CreateSavedPromptRequest"
      responses:
        "201":
          description: Saved prompt created
          content:
            application/json:
              schema:
                $ref: "./schemas/cambot/saved-prompt.yaml#/components/schemas/SavedPrompt"
  /saved-prompts/{promptId}:
    get:
      tags: [Saved Prompts]
      summary: Get a saved prompt
      operationId: getSavedPrompt
      parameters:
        - $ref: "#/components/parameters/PromptId"
      responses:
        "200":
          description: Saved prompt returned
          content:
            application/json:
              schema:
                $ref: "./schemas/cambot/saved-prompt.yaml#/components/schemas/SavedPrompt"

    put:
      tags: [Saved Prompts]
      summary: Update a saved prompt
      operationId: updateSavedPrompt
      parameters:
        - $ref: "#/components/parameters/PromptId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/cambot/saved-prompt.yaml#/components/schemas/UpdateSavedPromptRequest"
      responses:
        "200":
          description: Saved prompt updated
          content:
            application/json:
              schema:
                $ref: "./schemas/cambot/saved-prompt.yaml#/components/schemas/SavedPrompt"

    delete:
      tags: [Saved Prompts]
      summary: Delete a saved prompt
      operationId: deleteSavedPrompt
      parameters:
        - $ref: "#/components/parameters/PromptId"
      responses:
        "204":
          description: Saved prompt deleted
  /camera-groups/{groupId}/prompt-bindings:
    get:
      tags: [Prompt Bindings]
      summary: List saved prompts applied to a camera group
      operationId: listCameraGroupPromptBindings
      parameters:
        - $ref: "#/components/parameters/GroupId"
      responses:
        "200":
          description: Prompt bindings returned
          content:
            application/json:
              schema:
                $ref: "./schemas/cambot/prompt-binding.yaml#/components/schemas/PromptBindingListResponse"

    post:
      tags: [Prompt Bindings]
      summary: Apply a saved prompt to a camera group
      operationId: createCameraGroupPromptBinding
      parameters:
        - $ref: "#/components/parameters/GroupId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/cambot/prompt-binding.yaml#/components/schemas/CreatePromptBindingRequest"
      responses:
        "201":
          description: Prompt binding created
          content:
            application/json:
              schema:
                $ref: "./schemas/cambot/prompt-binding.yaml#/components/schemas/PromptBinding"
  /camera-groups/{groupId}/prompt-bindings/{bindingId}:
    put:
      tags: [Prompt Bindings]
      summary: Update a prompt binding
      operationId: updateCameraGroupPromptBinding
      parameters:
        - $ref: "#/components/parameters/GroupId"
        - $ref: "#/components/parameters/BindingId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/cambot/prompt-binding.yaml#/components/schemas/UpdatePromptBindingRequest"
      responses:
        "200":
          description: Prompt binding updated
          content:
            application/json:
              schema:
                $ref: "./schemas/cambot/prompt-binding.yaml#/components/schemas/PromptBinding"

    delete:
      tags: [Prompt Bindings]
      summary: Remove prompt from camera group
      operationId: deleteCameraGroupPromptBinding
      parameters:
        - $ref: "#/components/parameters/GroupId"
        - $ref: "#/components/parameters/BindingId"
      responses:
        "204":
          description: Prompt binding deleted
  /operations/estimate:
    post:
      tags: [Operations]
      summary: Estimate prompt scan usage and cost
      operationId: estimateOperation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/cambot/operation.yaml#/components/schemas/OperationEstimateRequest"
      responses:
        "200":
          description: Operation estimate returned
          content:
            application/json:
              schema:
                $ref: "./schemas/cambot/operation.yaml#/components/schemas/OperationEstimate"
  /operations:
    get:
      tags: [Operations]
      summary: List prompt scan operations
      operationId: listOperations
      parameters:
        - name: promptId
          in: query
          required: false
          schema:
            type: string
        - name: cameraGroupId
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            $ref: "./schemas/cambot/operation.yaml#/components/schemas/OperationStatus"
      responses:
        "200":
          description: Operations returned
          content:
            application/json:
              schema:
                $ref: "./schemas/cambot/operation.yaml#/components/schemas/OperationListResponse"

    post:
      tags: [Operations]
      summary: Create a prompt scan operation
      operationId: createOperation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/cambot/operation.yaml#/components/schemas/CreateOperationRequest"
      responses:
        "201":
          description: Operation created
          content:
            application/json:
              schema:
                $ref: "./schemas/cambot/operation.yaml#/components/schemas/Operation"
  /operations/{operationId}:
    get:
      tags: [Operations]
      summary: Get a prompt scan operation
      operationId: getOperation
      parameters:
        - name: operationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Operation returned
          content:
            application/json:
              schema:
                $ref: "./schemas/cambot/operation.yaml#/components/schemas/Operation"
  /operations/{operationId}/first-pass-results:
    get:
      tags: [Operations]
      summary: List first-pass image results for an operation
      operationId: listOperationFirstPassResults
      parameters:
        - name: operationId
          in: path
          required: true
          schema:
            type: string
        - name: include
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: First-pass results returned
          content:
            application/json:
              schema:
                $ref: "./schemas/cambot/operation.yaml#/components/schemas/FirstPassResultListResponse"
  /operations/{operationId}/second-pass-results:
    get:
      tags: [Operations]
      summary: List second-pass global prompt results for an operation
      operationId: listOperationSecondPassResults
      parameters:
        - name: operationId
          in: path
          required: true
          schema:
            type: string
        - name: include
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Second-pass results returned
          content:
            application/json:
              schema:
                $ref: "./schemas/cambot/operation.yaml#/components/schemas/SecondPassResultListResponse"
  /prompt-results/latest/first-pass:
    get:
      tags: [Operations]
      summary: List latest first-pass results for a prompt and camera group
      operationId: listLatestFirstPassResults
      parameters:
        - name: promptId
          in: query
          required: true
          schema:
            type: string
        - name: cameraGroupId
          in: query
          required: true
          schema:
            type: string
        - name: include
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Latest first-pass results returned
          content:
            application/json:
              schema:
                $ref: "./schemas/cambot/operation.yaml#/components/schemas/LatestFirstPassResultListResponse"
  /prompt-results/latest/second-pass:
    get:
      tags: [Operations]
      summary: List latest second-pass global results for a prompt and camera group
      operationId: listLatestSecondPassResults
      parameters:
        - name: promptId
          in: query
          required: true
          schema:
            type: string
        - name: cameraGroupId
          in: query
          required: true
          schema:
            type: string
        - name: include
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Latest second-pass results returned
          content:
            application/json:
              schema:
                $ref: "./schemas/cambot/operation.yaml#/components/schemas/LatestSecondPassResultListResponse"
  /operator-queue:
    get:
      tags: [Operator Queue]
      summary: List operator queue items
      operationId: listOperatorQueueItems
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "./schemas/cambot/operator-queue.yaml#/components/schemas/OperatorQueueStatus"
      responses:
        "200":
          description: Operator queue returned
          content:
            application/json:
              schema:
                $ref: "./schemas/cambot/operator-queue.yaml#/components/schemas/OperatorQueueListResponse"

    post:
      tags: [Operator Queue]
      summary: Create an operator queue item from a second-pass result
      operationId: createOperatorQueueItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/cambot/operator-queue.yaml#/components/schemas/CreateOperatorQueueItemRequest"
      responses:
        "201":
          description: Queue item created
          content:
            application/json:
              schema:
                $ref: "./schemas/cambot/operator-queue.yaml#/components/schemas/OperatorQueueItem"
  /operator-queue/{queueItemId}:
    put:
      tags: [Operator Queue]
      summary: Update operator queue item status
      operationId: updateOperatorQueueItem
      parameters:
        - name: queueItemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/cambot/operator-queue.yaml#/components/schemas/UpdateOperatorQueueItemRequest"
      responses:
        "200":
          description: Queue item updated
          content:
            application/json:
              schema:
                $ref: "./schemas/cambot/operator-queue.yaml#/components/schemas/OperatorQueueItem"
  /settings/gemini:
    get:
      tags: [Settings]
      summary: Get Gemini caller settings
      operationId: getGeminiCallerSettings
      responses:
        "200":
          description: Gemini caller settings returned
          content:
            application/json:
              schema:
                $ref: "./schemas/cambot/settings.yaml#/components/schemas/GeminiCallerSettings"

    put:
      tags: [Settings]
      summary: Update Gemini caller settings
      operationId: updateGeminiCallerSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/cambot/settings.yaml#/components/schemas/UpdateGeminiCallerSettingsRequest"
      responses:
        "200":
          description: Gemini caller settings updated
          content:
            application/json:
              schema:
                $ref: "./schemas/cambot/settings.yaml#/components/schemas/GeminiCallerSettings"
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: HTTP Basic authentication using the operator username and password. Send credentials with the Authorization header.
  parameters:
    GroupId:
      name: groupId
      in: path
      required: true
      schema:
        type: string
    PromptId:
      name: promptId
      in: path
      required: true
      schema:
        type: string
    BindingId:
      name: bindingId
      in: path
      required: true
      schema:
        type: string